requirements.txt
openai>=1.40.0
google-genai>=1.0.0
python-docx>=1.1.0
pillow>=10.0.0
google-cloud-speech>=2.26.0
